Add support for end of the month reminders
We want to remind developers to submit timesheets on the last day of the month.
This solution assumes that timesheets are submitted on each Friday and prompts for
submission for the missing hours on the last week of the month from Monday till the last day of the month(midweek).
The solution assumes that a part-time developer has their non-worked day at the beginning of the week.
No end of the month reminders are sent out on Fridays or during weekend.