Add test UserAssignments data ready for stubbing
The Harvest API has an endpoint for `UserAssignments`, a data structure
that links Users with Projects.

Hitting the UserAssignments endpoint is necessary to determine which
Project Manager is associated with which Project.

This commit adds dummy data ready for the Harvest API stub to respond
requests to the UserAssignments endpoint.