Link ProjectManager to TimeEntry in HarvestGateway
Write CanAssociateProjectManagersWithTimeSheets() unit test.

Define a private method GetProjectManagerId to hit the Harvest API
endpoint for a given project and get the associated ProjectManagerId.

GetProjectManagerId is then used in the RetrieveTimeSheets public
method to set the TimeSheet attribute ProjectManagerId.

Unit test asserts that expected ProjectManagerId is returned
with the TimeSheet.