Implement GetProjectManagers... Execute() method
This commit completes the functionality of
GetProjectManagersWithOpenTimeEntries.

To note:
1) To avoid changing the interface, the method returns a
GetLateBillablePeopleResponse.  We may want to modify
the interface to a higher level of abstraction later (as the concern is
Project Managers, not LateBillablePeople), but I didn't want
to change too much with this PR.

2) Several private methods are straight copies from
GetLateBillablePeople (e.g. formatting of email addresses for Slack
and date calculation).  We may want to extract these later.

3) I found it difficult to test this object.  In particular, I struggled
to get my head around the various types in play.  There are no
unit tests.  We may decide that the acceptance test is sufficient.  (I'm
very open to help/feedback/ideas from others on how to test this.)