Add feature flag to RemindProjectManagers
This commit sets the default behaviour of RemindProjectManagers
to only send a Slack message to the developer of the feature (Zack)
instead of to the Project Manager directly.

This is to fully confirm correct behaviour in production by allowing
comparison of which Project Managers the app would have reminded
vs which Project Managers _should_ have been reminded based on
Harvest.

Once we are confident in the feature, this flag can be removed.