Add retry policy to HarvestGateway request method
This commit uses the fault-handling library Polly to
retry requests to the Harvest API if they are rejected with an HTTP
status code "429 - Too Many Requests" (i.e. if the rate limit is
exceeded).

It does this using exponential back-off (`(Math.Pow(2, attempt)`)
to increase the time between retried requests, configurable with the
constants THROTTLE_TIME_IN_MS and MAX_ATTEMPTS.