Remove references to harvest in acceptance tests
The Wolves system will no longer integrate with harvest so I'm removing
all references to it in the acceptance tests. The goal is to make The Wolves
send a reminder to everyone at specific times and dates, since we can't
track timesheet completion programmatically anymore due to our migration
to Kimble.